requirements.txt
streamlit
python-dotenv
langchain-groq
langchain-core
langchain-ollama
st-copy-to-clipboard
